flare: give the source mip back a way in, as the integer it is
The review round removed its slider for being a fractional control over what is
an integer index, which was right about the control and wrong about the remedy:
it left a tuning parameter with no way to reach it at all. No CLI flag, no scene
key, no GUI. The other flare knobs are GUI-only too, so restoring the slider
puts this back in line with its siblings rather than inventing a one-off flag
for the one parameter that lost its own.

The field is now an int, because that is what a mip index is. textureLod takes
a float and would happily blend between levels, but nothing wants a ghost read
half out of one mip and half out of the next, and a float invited exactly the
fractional slider that got deleted.

Its top is the pyramid that exists at this resolution, not a constant. The mip
count falls out of the frame size, so a fixed ceiling would let the slider
travel past the last real level and do nothing -- the same class of dead-end as
a halo width whose zero was its brightest setting.

Clamp arithmetic is unchanged, so both goldens still measure 0 px and the gates
report identical numbers.

+++ cetra/src/postfx.h
@@ -212,7 +212,7 @@ typedef struct PostFX {
     float flare_ghost_spacing;       // Fraction of the centre vector between ghosts
     float flare_halo_width;          // Halo ring radius in UV units
     float flare_chroma;              // Per-channel radial offset; 0 = achromatic
-    float flare_source_lod;          // Pyramid level; a mid mip, clamped to what exists
+    int flare_source_lod;            // Pyramid mip the ghosts read, clamped to what exists
     bool flare_ready;                // Lazy-alloc guard for the target below
     GLuint flare_fbo, flare_texture; // Quarter post-res
     int flare_width, flare_height;
